Move IdentifyScrub and IdentifyCow to randomizer_identify_actor.cpp
