+++ soh/src/overlays/actors/ovl_En_Cow/z_en_cow.h
@@ -21,9 +21,4 @@ typedef struct EnCow {
     /* 0x027C */ EnCowActionFunc actionFunc;
 } EnCow; // size = 0x0280
 
-typedef struct CowInfo {
-    RandomizerInf randomizerInf;
-    RandomizerCheck randomizerCheck;
-} CowInfo;
-
 #endif
